Added icon and placeholder photo. Consolidated 3 databases to one database with 3 tables. Updated save/update logic.

+++ TripJournal/Place.m
@@ -10,7 +10,7 @@
 
 @implementation Place
 
-- (id)initWithUniqueId:(int)uniqueId tripId:(NSNumber *)tripId name:(NSString *)name description:(NSString *)description photo:(NSString *)photo startDate:(NSDate *)startDate endDate:(NSDate *)endDate {
+- (id)initWithUniqueId:(long long)uniqueId tripId:(NSNumber *)tripId name:(NSString *)name description:(NSString *)description photo:(NSString *)photo startDate:(NSDate *)startDate endDate:(NSDate *)endDate {
     if ((self = [super init])) {
         self.uniqueId = uniqueId;
         self.tripId = tripId;
